requirements.txt
googletrans==3.0.0
langchain==0.2.13
pandas==1.5.3
pymongo==4.3.3
PyPDF2==3.0.1
python_docx==1.1.2
streamlit==1.18.1
